+++ src/org/mixare/LocalMarker.java
@@ -177,7 +177,7 @@ public abstract class LocalMarker implements Marker {
 		//TODO adapt the following to the variable radius!
 		pPt.x = x - signMarker.x;
 		pPt.y = y - signMarker.y;
-		pPt.rotate(Math.toRadians(-(currentAngle + 90)));
+		pPt.rotate((float) Math.toRadians(-(currentAngle + 90)));
 		pPt.x += txtLab.getX();
 		pPt.y += txtLab.getY();
